leetcode/0212.Word-Search-II/212.WordSearchII_test.go
package leetcode

import (
	"reflect"
	"testing"
)

func Test_findWords(t *testing.T) {
	type args struct {
		board [][]byte
		words []string
	}
	tests := []struct {
		name string
		args args
		want []string
	}{
		{
			name: `board = [["o","a","a","n"]...], words = ["oath",...]`,
			args: args{[][]byte{{'o', 'a', 'a', 'n'}, {'e', 't', 'a', 'e'}, {'i', 'h', 'k', 'r'}, {'i', 'f', 'l', 'v'}}, []string{"oath", "pea", "eat", "rain"}},
			want: []string{"oath", "eat"},
		},
		{
			name: `board = [["o","a","a","n"]...], words = ["oate",...]`,
			args: args{[][]byte{{'o', 'a', 'a', 'n'}, {'e', 't', 'a', 'e'}, {'i', 'h', 'k', 'r'}, {'i', 'f', 'l', 'v'}}, []string{"oath", "pea", "eat", "rain", "oathk", "oathf", "oathfi", "oathfii", "oathi", "oathii", "oate"}},
			want: []string{"oath", "oathf", "oathfi", "oathfii", "oathk", "oathi", "oathii", "oate", "eat"},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findWords(tt.args.board, tt.args.words); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("findWords() = %v, want %v", got, tt.want)
			}
		})
	}
}
